Some cleanup
Some file name and content hygiene.
