+++ pkg/placement/dynamic-join.go
