+++ pkg/placement/dynamic-join_test.go
